Added additional development info, fix Bower deps.
After running into some trouble building the application for
development, I found that Bower had to be installed globally.

I added that information to the development section to the README
so that other developers attempting to build the project will
avoid problems.

There also was an issue with the bower.json file that prevented
the bower install to succeed.  I changed the Bootstrap dependency
to point to the latest 3.2.* release, which fixes the build.